+++ drivers/video/video_buffer.h
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2025 NXP
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+struct video_buffer *video_get_buf_sqe(struct mpsc *io_q);
